+++ console/packages/starwhale-ui/src/TiptapEditor/components/EditorBubbleMenu.tsx
@@ -1,18 +1,11 @@
 import { BubbleMenu, BubbleMenuProps } from '@tiptap/react'
 import { FC, useState } from 'react'
 import { BoldIcon, ItalicIcon, UnderlineIcon, StrikethroughIcon, CodeIcon } from 'lucide-react'
-//  eslint-disable-next-line import/no-cycle
 import { NodeSelector } from './node-selector'
 import { ColorSelector } from './color-selector'
 import { LinkSelector } from './link-selector'
 import { cn } from '../lib/utils'
-
-export interface BubbleMenuItem {
-    name: string
-    isActive: () => boolean
-    command: () => void
-    icon: typeof BoldIcon
-}
+import { BubbleMenuItem } from './types'
 
 type EditorBubbleMenuProps = Omit<BubbleMenuProps, 'children'>
