+++ console/packages/starwhale-ui/src/TiptapEditor/components/node-selector.tsx
@@ -12,8 +12,7 @@ import {
     CheckSquare,
 } from 'lucide-react'
 import { Dispatch, FC, SetStateAction } from 'react'
-// eslint-disable-next-line import/no-cycle
-import { BubbleMenuItem } from './EditorBubbleMenu'
+import { BubbleMenuItem } from './types'
 
 interface NodeSelectorProps {
     editor: Editor
