+++ console/packages/starwhale-ui/src/TiptapEditor/components/types.ts
@@ -1,0 +1,8 @@
+import { BoldIcon } from 'lucide-react'
+
+export interface BubbleMenuItem {
+    name: string
+    isActive: () => boolean
+    command: () => void
+    icon: typeof BoldIcon
+}
